+++ NEWS.md
@@ -1,11 +1,21 @@
-# tacmagic 0.2.1
+# tacmagic News
 
-## Minor changes
+## tacmagic 0.3.0 (in development)
+
+### Major changes
+
+- WIP Radioactivity unit conversion can now be done for tac objects and numeric objects
+- WIP Conversion of a tac object to standardized uptake value (SUV) by controlling for weight and injected dose
+- TODO Vignette updated with new features
+
+## tacmagic 0.2.1 (2019-03-06)
+
+### Minor changes
 
 This version has no changes to features but addresses minor, mostly stylistic, issues for the initial CRAN release.
 
-# tacmagic 0.2.0
+## tacmagic 0.2.0 (2019-02-26)
 
-## Major changes
+### Major changes
 
 This is the first release as a complete R package, and the version that has undergone open review with rOpenSci. The main features of tacmagic 0.2.0 are implemented and tested: loading PET time activity curve (tac) data from multiple formats, merging ROIs weighted for volume, calculating binding potential models including SUVR and DVR, basic plotting, and calculation of cut-off values. As tacmagic is still a new package and testing to date is based on few example files, it is essential to verify that all functions are behaving as expected when applied to your own data.